drop redundant static_cast<int> on items_for_vec
Per @fbusato: ceil_div's result initializes a constexpr int directly; the
size_t->int copy-init is well-formed (not list-init) and CCCL's warning set
(-Wall -Wextra, no -Wconversion) doesn't flag it, so the cast was noise.

+++ cub/cub/device/dispatch/tuning/tuning_transform_tile.cuh
@@ -50,7 +50,7 @@ constexpr int pick_tile_size(bool mufu_heavy = false, ::cuda::compute_capability
   constexpr int max_occupancy        = 16;
 
   constexpr auto min_elem     = ::cuda::std::min({sizeof(Out), sizeof(Ins)...});
-  constexpr int items_for_vec = static_cast<int>(::cuda::ceil_div(vector_bytes, min_elem));
+  constexpr int items_for_vec = ::cuda::ceil_div(vector_bytes, min_elem);
 
   // Fill (zero inputs) keeps the same latency target by counting output bytes.
   constexpr auto bytes_per_iter = (sizeof...(Ins) > 0) ? (sizeof(Ins) + ... + ::cuda::std::size_t{0}) : sizeof(Out);
